get data in form write will send by axios

+++ server/controllers/post.controller.js
@@ -27,14 +27,19 @@ export const createPost = async (req, res) => {
         return res.status(404).json("User not found!");
     }
 
-
-
     const newPost = new Post({ user: user._id, ...req.body });
     const post = await newPost.save();
     res.status(200).json(post);
 };
 
 export const deletePost = async (req, res) => {
-    const post = await Post.findOneAndDelete({ _id: req.params.id, user: user._id });
+    const clerkUserId = req.auth.userId;
+    if (!clerkUserId) {
+        return res.status(401).json("Not authenticated!");
+    }
+    const user = await User.findOne({ clerkUserId });
+
+    const deletePost = await Post.findOneAndDelete({ _id: req.params.id, user: user._id });
     res.status(200).json("post has deleted");
+
 };
